Revert "silently delete tmp file instead of asking confirmation"
This reverts commit 8e96bb5c3601e58cca866b2a9f9a361fd98b2593.

Conflicts:
	bin/install.sh

+++ bin/install.sh
@@ -276,7 +276,9 @@ cd ${TEMP_DIR}
 
 # delete existing tmp file before download again
 if [[ -e  ${PIO_FILE} ]]; then
-  rm ${PIO_FILE}
+  if confirm "Delete existing temporary file?"; then
+    rm ${PIO_FILE}
+  fi
 fi
 
 if [[ ! -e ${PIO_FILE} ]]; then
